Backend/tests/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-auth==2.38.0
google-genai==1.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==15.0.1
